Starting Progress on Website
Progress includes a header with bkg and name, and side nav that expands out on hamburger icon click for small/ medium screens. Nav is static on screens larger than a medium size. 
Scratch paper includes content that will go into the site in the future.
Color scheme has not yet been added to the design.